[vuex] restructure store modules
